+++ homeassistant/components/lutron/fan.py
@@ -10,8 +10,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import DOMAIN, LutronController, LutronData
-from .aiolip import Output
+from . import DOMAIN, LutronData
+from .aiolip import LutronController, Output
 from .entity import LutronOutput
 
 _LOGGER = logging.getLogger(__name__)
