+++ homeassistant/components/lutron/switch.py
@@ -10,8 +10,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import CONF_USE_RADIORA_MODE, DOMAIN, LIPLedState, LutronController, LutronData
-from .aiolip import Led
+from . import CONF_USE_RADIORA_MODE, DOMAIN, LutronData
+from .aiolip import Led, LIPLedState, LutronController
 from .entity import LutronKeypadComponent, LutronOutput
